requirements.txt
altair==4.2.0rc1
aniso8601==9.0.1
argcomplete==1.12.3
astor==0.8.1
astroid==2.6.6
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.12b0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
flask-restplus==0.13.0
fonttools==4.28.3
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==8.0.0b0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-widgets==2.0.0b0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.23
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pydeck==0.7.1
Pygments==2.10.0
pylint==3.0.0a4
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.27
streamlit==1.2.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.4.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==2.0.2
widgetsnbextension==4.0.0b0
wrapt==1.12.1
zipp==3.6.0
streamlit
